RYA-104 Updated the RyaClient to optionally include features that are required for the Accumulo Client, but not the Mongo Client.
