+++ extras/indexing/src/main/java/org/apache/rya/api/client/mongo/MongoRyaClientFactory.java
@@ -70,5 +70,5 @@ public class MongoRyaClientFactory {
                 new MongoLoadStatements(connectionDetails, instanceExists),
                 new MongoLoadStatementsFile(connectionDetails, instanceExists),
                 new MongoExecuteSparqlQuery(connectionDetails, instanceExists));
-    }
+   }
 }
