+++ extras/rya.benchmark/src/test/java/org/apache/rya/benchmark/query/QueryBenchmarkRunIT.java
@@ -148,7 +148,7 @@ public class QueryBenchmarkRunIT {
             final String pcjId = pcjs.createPcj(SPARQL_QUERY);
 
             // Batch update the PCJ using the Rya Client.
-            ryaClient.getBatchUpdatePCJ().batchUpdate(RYA_INSTANCE_NAME, pcjId);
+            ryaClient.getBatchUpdatePCJ().get().batchUpdate(RYA_INSTANCE_NAME, pcjId);
         }
     }
