+++ extras/rya.pcj.fluo/pcj.fluo.integration/src/test/java/org/apache/rya/indexing/pcj/fluo/integration/QueryIT.java
@@ -996,7 +996,7 @@ public class QueryIT extends RyaExportITBase {
 
         switch (strategy) {
         case RYA:
-            ryaClient.getCreatePCJ().createPCJ(getRyaInstanceName(), sparql);
+            ryaClient.getCreatePCJ().get().createPCJ(getRyaInstanceName(), sparql);
             addStatementsAndWait(statements);
             // Fetch the value that is stored within the PCJ table.
             try (final PrecomputedJoinStorage pcjStorage = new AccumuloPcjStorage(accumuloConn, getRyaInstanceName())) {
